+++ .ci/containers/membership-checker/utils_test.go
@@ -3,25 +3,7 @@ package main
 import (
 	"reflect"
 	"testing"
-
-	"golang.org/x/exp/slices"
 )
-
-func TestTrustedContributors(t *testing.T) {
-	for _, member := range trustedContributors {
-		if slices.Contains(reviewerRotation, member) {
-			t.Fatalf(`%v should not be on reviewerRotation list`, member)
-		}
-	}
-}
-
-func TestOnVacationReviewers(t *testing.T) {
-	for _, member := range onVacationReviewers {
-		if !slices.Contains(reviewerRotation, member) {
-			t.Fatalf(`%v is not on reviewerRotation list`, member)
-		}
-	}
-}
 
 func TestRemovesList(t *testing.T) {
 	cases := map[string]struct {
