+++ .ci/gcb-generate-diffs-new.yml
@@ -106,6 +106,7 @@ steps:
 
     - name: 'gcr.io/graphite-docker-images/downstream-builder'
       id: tfv-head
+      allowFailure: true
       secretEnv: ["GITHUB_TOKEN"]
       waitFor: ["merged", "tpg-head"]
       args:
@@ -116,6 +117,7 @@ steps:
 
     - name: 'gcr.io/graphite-docker-images/downstream-builder'
       id: tfv-base
+      allowFailure: true
       secretEnv: ["GITHUB_TOKEN"]
       waitFor: ["merged", "tpg-base"]
       args:
@@ -156,6 +158,7 @@ steps:
 
     - name: 'gcr.io/graphite-docker-images/terraform-validator-tester'
       id: tfv-test
+      allowFailure: true
       secretEnv: ["GITHUB_TOKEN"]
       waitFor: ["tfv-head", "tfv-base"]
       env:
@@ -172,6 +175,7 @@ steps:
 
     - name: 'gcr.io/graphite-docker-images/terraform-validator-tester-integration'
       id: tfv-test-integration-0.12.31
+      allowFailure: true
       secretEnv: ["GITHUB_TOKEN"]
       waitFor: ["tfv-head", "tfv-base"]
       env:
@@ -189,6 +193,7 @@ steps:
 
     - name: 'gcr.io/graphite-docker-images/terraform-validator-tester-integration'
       id: tfv-test-integration-0.13.7
+      allowFailure: true
       secretEnv: ["GITHUB_TOKEN"]
       waitFor: ["tfv-head", "tfv-base"]
       env:
